requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.4.0
blinker==1.4              # via pelican
click==5.1                # via pip-tools
docutils==0.12            # via pelican
feedgenerator==1.7        # via pelican
first==2.0.1              # via pip-tools
htmlmin==0.1.6            # via pelican-minify
hurry.filesize==0.9
jinja2==2.8               # via pelican
markdown==2.6.2
markupsafe==0.23          # via jinja2
pelican-minify==0.8
pelican==3.6.3
pip-tools==1.1.3
pygments==2.0.2           # via pelican
python-dateutil==2.4.2    # via pelican
pytz==2015.4              # via feedgenerator, pelican
requests==2.7.0
six==1.9.0                # via feedgenerator, pelican, pip-tools, python-dateutil
unidecode==0.4.18         # via pelican

# The following packages are commented out because they are
# considered to be unsafe in a requirements file:
# setuptools==18.2          # via hurry.filesize
